// external/@wdk/account-abstraction-ton/packages/ui-stargate-usdv/src/typechain/USDVMain.ts
export type MessagingFeeStruct = {
  nativeFee: PromiseOrValue<BigNumberish>;
  lzTokenFee: PromiseOrValue<BigNumberish>;
};
// ...
export declare namespace IOFT {
  export type SendParamStruct = {
    to: PromiseOrValue<BytesLike>;
    amountLD: PromiseOrValue<BigNumberish>;
    minAmountLD: PromiseOrValue<BigNumberish>;
    dstEid: PromiseOrValue<BigNumberish>;
  };

  export type SendParamStructOutput = [string, BigNumber, BigNumber, number] & {
    to: string;
    amountLD: BigNumber;
    minAmountLD: BigNumber;
    dstEid: number;
  };
}